Deprecated SXNET, some old NETSCAPE data types
Support for the SXNET and (Thawte Strong Extranet) certificate extension,
the NETSCAPE_CERT_SEQUENCE, NETSCAPE_SPKAC, and NETSCAPE_SPKI datatypes
have been deprecrated.

Fixes #1177, #10668
