+++ apps/build.info
@@ -14,10 +14,10 @@ $OPENSSLSRC=\
         openssl.c \
         asn1parse.c ca.c ciphers.c crl.c crl2pkcs7.c dgst.c \
         enc.c errstr.c \
-        genpkey.c kdf.c mac.c nseq.c passwd.c pkcs7.c \
+        genpkey.c kdf.c mac.c passwd.c pkcs7.c \
         pkcs8.c pkey.c pkeyparam.c pkeyutl.c prime.c rand.c req.c \
         s_client.c s_server.c s_time.c sess_id.c smime.c speed.c \
-        spkac.c verify.c version.c x509.c rehash.c storeutl.c \
+        verify.c version.c x509.c rehash.c storeutl.c \
         list.c info.c fipsinstall.c pkcs12.c
 IF[{- !$disabled{'ec'} -}]
   $OPENSSLSRC=$OPENSSLSRC ec.c ecparam.c
@@ -53,6 +53,9 @@ IF[{- !$disabled{'cms'} -}]
 ENDIF
 IF[{- !$disabled{'cmp'} -}]
   $OPENSSLSRC=$OPENSSLSRC cmp.c lib/cmp_mock_srv.c
+ENDIF
+IF[{- !$disabled{'deprecated-3.0'} -}]
+  $OPENSSLSRC=$OPENSSLSRC spkac.c nseq.c
 ENDIF
 
 IF[{- !$disabled{apps} -}]
