+++ crypto/asn1/asn1_item_list.h
@@ -84,9 +84,11 @@ static ASN1_ITEM_EXP *asn1_item_list[] = {
     ASN1_ITEM_ref(LONG),
 #endif
     ASN1_ITEM_ref(NAME_CONSTRAINTS),
+#ifndef OPENSSL_NO_DEPRECATED_3_0
     ASN1_ITEM_ref(NETSCAPE_CERT_SEQUENCE),
     ASN1_ITEM_ref(NETSCAPE_SPKAC),
     ASN1_ITEM_ref(NETSCAPE_SPKI),
+#endif
     ASN1_ITEM_ref(NOTICEREF),
 #ifndef OPENSSL_NO_OCSP
     ASN1_ITEM_ref(OCSP_BASICRESP),
@@ -145,8 +147,10 @@ static ASN1_ITEM_EXP *asn1_item_list[] = {
 #ifndef OPENSSL_NO_SCRYPT
     ASN1_ITEM_ref(SCRYPT_PARAMS),
 #endif
+#ifndef OPENSSL_NO_DEPRECATED_3_0
     ASN1_ITEM_ref(SXNETID),
     ASN1_ITEM_ref(SXNET),
+#endif
     ASN1_ITEM_ref(ISSUER_SIGN_TOOL),
     ASN1_ITEM_ref(USERNOTICE),
     ASN1_ITEM_ref(X509_ALGORS),
