+++ crypto/asn1/build.info
@@ -4,9 +4,9 @@ SOURCE[../../libcrypto]=\
         a_print.c a_type.c a_dup.c a_d2i_fp.c a_i2d_fp.c \
         a_utf8.c a_sign.c a_digest.c a_verify.c a_mbstr.c a_strex.c \
         x_algor.c x_val.c x_sig.c x_bignum.c \
-        x_int64.c x_info.c x_spki.c nsseq.c \
+        x_int64.c x_info.c \
         d2i_pu.c d2i_pr.c i2d_evp.c \
-        t_pkey.c t_spki.c t_bitst.c \
+        t_pkey.c t_bitst.c \
         tasn_new.c tasn_fre.c tasn_enc.c tasn_dec.c tasn_utl.c tasn_typ.c \
         tasn_prn.c tasn_scn.c ameth_lib.c \
         f_int.c f_string.c \
@@ -19,5 +19,5 @@ IF[{- !$disabled{'rsa'} and !$disabled{'rc4'} -}]
   SOURCE[../../libcrypto]=n_pkey.c
 ENDIF
 IF[{- !$disabled{'deprecated-3.0'} -}]
-  SOURCE[../../libcrypto]=x_long.c
+  SOURCE[../../libcrypto]=x_long.c x_spki.c nsseq.c t_spki.c
 ENDIF
