+++ crypto/x509/standard_exts.h
@@ -37,7 +37,9 @@ static const X509V3_EXT_METHOD *standard_exts[] = {
 #ifndef OPENSSL_NO_OCSP
     &ossl_v3_crl_invdate,
 #endif
+#ifndef OPENSSL_NO_DEPRECATED_3_0
     &ossl_v3_sxnet,
+#endif
     &ossl_v3_info,
 #ifndef OPENSSL_NO_RFC3779
     &ossl_v3_addr,
